+++ ctrlf.el
@@ -976,7 +976,9 @@ I have literally no idea why this is needed.")
 And self-destruct this hook."
   (remove-hook 'post-command-hook #'ctrlf--finalize)
   (unless (= (point) ctrlf--starting-point)
-    (push-mark ctrlf--starting-point)))
+    (if ctrlf-auto-recenter
+        (set-window-start (get-buffer-window) ctrlf--final-window-start)
+      (push-mark ctrlf--starting-point))))
 
 (defun ctrlf--minibuffer-exit-hook ()
   "Clean up CTRLF from minibuffer and self-destruct this hook."
@@ -1490,7 +1492,9 @@ search, change back to fuzzy-regexp search."
     "Minor mode to use CTRLF in place of Isearch."
     :keymap ctrlf-mode-map
     (require 'map)
-    (let ((default-ctrlf-mode-bindings
+    ;; Weird indentation to make things indent the same in both Emacs
+    ;; 28 and Emacs 29, because apparently something changed?
+    (let (( default-ctrlf-mode-bindings
             (eval (car (get 'ctrlf-mode-bindings 'standard-value)))))
       (when (and ctrlf-local-mode
                  default-ctrlf-mode-bindings